python_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pymongo==4.6.0
youtube-transcript-api==0.6.2
youtube-search-python==1.6.6
yt-dlp==2023.12.30
requests==2.31.0
google-generativeai==0.3.2
python-multipart==0.0.6
google-cloud-aiplatform==1.71.1
google-cloud-functions==1.16.3
vertexai==1.71.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
redis==5.0.1
apscheduler==3.10.4
sentence-transformers==2.2.2
python-dotenv==1.0.0
transformers==4.35.2
tensorflow==2.15.0
torch==2.1.1
kaggle==1.5.16 